requirements.txt
streamlit
pandas
openpyxl
Pillow
random
os
